move tf related staff to tf_utils.py, since tf is in requiremnets, get rid of check_tensorflow
